repo reorg: modularize caltools.py
