+++ src/nisarqa/processing/__init__.py
@@ -1,6 +1,7 @@
 from .plotting_utils import *
 from .processing_utils import *
 
+from .caltools import *
 from .setup_pdf import *
 from .calc import *
 from .multilook import *
