+++ src/nisarqa/products/__init__.py
@@ -1,2 +1,2 @@
 # keep individual products in their own namespace
-from . import caltools, gcov, gslc, igram, offsets, rslc
+from . import gcov, gslc, igram, offsets, rslc
